quizform - impl different types
