+++ harusbisa-frontend/src/css/appDefault.css
@@ -35,7 +35,6 @@ p, label{
 input:focus{
     outline-color: var(--hb-yellow);
     outline-width: 1px;
-    border-radius: 8px;
 }
 
 /*BUTTON*/
